// src/types/api.ts
export interface components {
    schemas: {
// ...
        ChatCompletion: {
            /** Id */
            id: string;
            /** Choices */
            choices: components["schemas"]["openai__types__chat__chat_completion__Choice"][];
            /** Created */
            created: number;
            /** Model */
            model: string;
            /**
             * Object
             * @constant
             */
            object: "chat.completion";
            /** Service Tier */
            service_tier?: ("auto" | "default" | "flex" | "scale" | "priority") | null;
            /** System Fingerprint */
            system_fingerprint?: string | null;
            usage?: components["schemas"]["CompletionUsage"] | null;
        } & {
            [key: string]: unknown;
        };
        /** ChatCompletionAssistantMessageParam */
        ChatCompletionAssistantMessageParam: {
            /**
             * Role
             * @constant
             */
            role: "assistant";
            audio?: components["schemas"]["Audio"] | null;
            /** Content */
            content?: string | (components["schemas"]["ChatCompletionContentPartTextParam"] | components["schemas"]["ChatCompletionContentPartRefusalParam"])[] | null;
            function_call?: components["schemas"]["speaches__types__chat__FunctionCall"] | null;
            /** Name */
            name?: string | null;
            /** Refusal */
            refusal?: string | null;
            /** Tool Calls */
            tool_calls?: components["schemas"]["ChatCompletionMessageToolCallParam"][] | null;
        };
// ...
        ChatCompletionSystemMessageParam: {
            /** Content */
            content: string | components["schemas"]["ChatCompletionContentPartTextParam"][];
            /**
             * Role
             * @constant
             */
            role: "system";
            /** Name */
            name?: string | null;
        };
// ...
        ChatCompletionUserMessageParam: {
            /** Content */
            content: string | (components["schemas"]["ChatCompletionContentPartTextParam"] | components["schemas"]["ChatCompletionContentPartImageParam"] | components["schemas"]["ChatCompletionContentPartInputAudioParam"])[];
            /**
             * Role
             * @constant
             */
            role: "user";
            /** Name */
            name?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}